Using method_missing to call allowed methods
Added a module business_centra/api_methods. This adds method_missing to the business_central/base class, which in turn is used to call the appropriate method.

When a subclass method is called, the SUPPORTED_METHODS array is checked to see if the method is supported for that Object. If it is, then the appropriate REST call is made.

Also using webmock to stub fixtures directly.